fix: Misuse of std::accumulate in TensorShape::collapse and total_size
TensorShape and Dimension classes's collapse and total_size calculations should have been using size_t or the template T , but the initial value used in std::accumulate is an integer, and it's always 32-bits. This prevents us calculating number of tensor elements correctly.

This also prevents us from reliably using ARM_COMPUTE_RETURN_ERROR_ON_SIZE_UNSUPPORTED(..) macro.

Other misuses of this call have also been corrected. but they are not bugs for library users.

Partially Resolves: ARMCL-1247
Relates to: COMPMID-8646

+++ src/core/utils/helpers/fft.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2020 Arm Limited.
+ * Copyright (c) 2019-2020, 2026 Arm Limited.
  *
  * SPDX-License-Identifier: MIT
  *
@@ -82,7 +82,7 @@ std::vector<unsigned int> digit_reverse_indices(unsigned int N, const std::vecto
 
     // Early exit in case N and fft stages do not match
     const float stages_prod =
-        std::accumulate(std::begin(fft_stages), std::end(fft_stages), 1, std::multiplies<unsigned int>());
+        std::accumulate(std::begin(fft_stages), std::end(fft_stages), (unsigned int)1, std::multiplies<unsigned int>());
     if (stages_prod != N)
     {
         return idx_digit_reverse;
